Move file helpers to Util
