+++ swagger/src/main/scala/me/apidoc/swagger/Parser.scala
@@ -2,11 +2,8 @@ package me.apidoc.swagger
 
 import lib.{ServiceConfiguration, Text, UrlKey}
 import java.io.File
-import java.nio.file.{Paths, Files}
-import java.nio.charset.StandardCharsets
 import scala.collection.JavaConversions._
 import play.api.libs.json.{Json, JsArray, JsObject, JsString, JsValue}
-import java.util.UUID
 
 import io.swagger.parser.SwaggerParser
 import com.wordnik.swagger.models.{ComposedModel, ModelImpl, RefModel, Swagger}
@@ -21,6 +18,12 @@ import scala.annotation.tailrec
 case class Parser(config: ServiceConfiguration) {
 
   private val DefaultResponseCode = "default"
+
+  def parseString(
+    contents: String
+  ): Service = {
+    parse(Util.writeToTempFile(contents))
+  }
 
   def parse(
     path: File
@@ -495,20 +498,6 @@ case class Parser(config: ServiceConfiguration) {
       val modelUrl = Converters.normalizeUrl(s"/${m.plural}")
       normalized == modelUrl || normalized.startsWith(modelUrl + "/")
     }
-  }
-
-  def parseString(
-    contents: String
-  ): Service = {
-    val tmpPath = "/tmp/apidoc.swagger.tmp." + UUID.randomUUID.toString + ".json"
-    writeToFile(tmpPath, contents)
-    parse(new File(tmpPath))
-  }
-
-  private def writeToFile(path: String, contents: String) {
-    val outputPath = Paths.get(path)
-    val bytes = contents.getBytes(StandardCharsets.UTF_8)
-    Files.write(outputPath, bytes)
   }
 
 }
